+++ R/terra_gdal_calc.R
@@ -24,11 +24,47 @@ NULL
 #'
 #' @inherit terra_gdal_project return
 #'
+#' @section Troubleshooting:
+#' The function aims to automatically determine the best approach to
+#' run the `gdal_calc.py` script. However, it is not always successful.
+#' As such, there are several environmental variables that can be
+#' configured to manually specify exactly how the `gdal_calc.py` script
+#' should be executed. These variables specify the following behavior.
+#'
+#' \describe{
+#'
+#' \item{\code{GDAL_PYTHON}}{Specifies the file path for the Python executable
+#'  used to run the `gdal_calc.py` script. For example, this variable
+#' could be configured as: `sys.setenv("GDAL_PYTHON", "python")`,
+#' `sys.setenv("GDAL_PYTHON", "python3")`, or
+#' `sys.setenv("GDAL_PYTHON", "python.exe")`.}
+#'
+#' \item{\code{GDAL_CALC}}{
+#' Specifies the file path for the \code{gdal_calc.py} script.
+#' For example, this variable could be configured as:
+#' `sys.setenv("GDAL_PYTHON", "gdal_calc.py")`, or
+#'` sys.setenv("GDAL_PYTHON",
+#' "C:\\OSGeo4W\\apps\\Python39\\Scripts\\gdal_calc.py")`.}
+#'
+#' \item{\code{GDAL_ESCAPE}}{
+#' Specifies whether symbols in the mathematical expressions used to perform the
+#' calculations should be escaped. On Windows systems, the default behavior
+#' is to escape these symbols. This behavior can be disabled using the
+#' following code: `sys.setenv("GDAL_ESCAPE", "false")`.}
+#'
+#' }
+#' @seealso
+#' See the package README for instructions to install the GDAL dependencies
+#' for this function. The [is_osgeo4w_available()] and
+#' [is_gdal_calc_available()] can be used to check if the installation
+#' was successful.
+#'
 #' @examples
 #' # please ensure that the Python and the GDAL system binaries are
-#' # installed to run the example
+#' # installed to run the example,
+#' # see ?is_gdal_calc_available for more details
 #'
-#' @examplesIf is_gdal_python_available()
+#' \dontrun{
 #' # create raster with data
 #' x <- rast(
 #'   ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
@@ -41,6 +77,7 @@ NULL
 #'
 #' # preview result
 #' print(y)
+#' }
 #' @export
 terra_gdal_calc <- function(x, expr,
                             y = NULL,
@@ -78,7 +115,7 @@ terra_gdal_calc <- function(x, expr,
     assertthat::noNA(output_raster),
     assertthat::is.string(datatype),
     assertthat::noNA(datatype),
-    is_gdal_python_available()
+    is_gdal_calc_available()
   )
   if (inherits(x, "SpatRaster")) {
     assertthat::assert_that(
@@ -95,6 +132,9 @@ terra_gdal_calc <- function(x, expr,
       terra::nlyr(z) == 1
     )
   }
+
+  # standardize output file path
+  filename <- normalize_path(filename, mustWork = FALSE)
 
   # compress options
   co <- paste0("NUM_THREADS=", n_threads)
@@ -141,8 +181,7 @@ terra_gdal_calc <- function(x, expr,
   }
 
   # build cmd processing
-  cmd <- "gdal_calc.py "
-  cmd <- paste0(cmd, "-X \"", f1, "\" ")
+  cmd <- paste0("-X \"", f1, "\" ")
   if (!is.null(y)) {
     cmd <- paste0(cmd, "-Y \"", f2, "\" ")
   }
@@ -168,6 +207,25 @@ terra_gdal_calc <- function(x, expr,
   if (!verbose) {
     cmd <- paste(cmd, "--quiet")
   }
+
+  # if on windows, we need to escape the characters: "(", ")", "<", "*"
+  # nocov start
+  if (
+    identical(.Platform$OS.type, "windows") &&
+    !identical(tolower(Sys.getenv("GDAL_ESCAPE")), "false")
+  ) {
+    cmd <- gsub("(", "^(", cmd, fixed = TRUE)
+    cmd <- gsub(")", "^)", cmd, fixed = TRUE)
+    cmd <- gsub("<", "^<", cmd, fixed = TRUE)
+    cmd <- gsub(">", "^<", cmd, fixed = TRUE)
+    cmd <- gsub("*", "^*", cmd, fixed = TRUE)
+    cmd <- gsub("|", "^|", cmd, fixed = TRUE)
+    cmd <- gsub("&", "^&", cmd, fixed = TRUE)
+  }
+  # nocov end
+
+  # prepend gdal_calc executable for calling script
+  cmd <- gdal_calc_command(cmd)
   if (isTRUE(verbose)) {
     cli::cli_alert_info(paste("System command:", cmd))
   }
@@ -194,4 +252,74 @@ terra_gdal_calc <- function(x, expr,
   } else {
     return(filename)
   }
+}
+
+gdal_calc_command <- function(x) {
+  if (
+    is_osgeo4w_available() &&
+    identical(Sys.getenv("GDAL_PYTHON"), "") &&
+    identical(Sys.getenv("GDAL_CALC"), "")
+  ) {
+    out <- osgeo4w_gdal_calc(x) # nocov
+  } else {
+    out <- python_gdal_calc(x)
+  }
+  out
+}
+
+# nocov start
+osgeo4w_gdal_calc <- function(x) {
+  # assert valid arguments
+  assertthat::assert_that(
+    assertthat::is.string(x),
+    assertthat::noNA(x)
+  )
+  # find OSGeo4W root
+  r <- Sys.getenv("OSGEO4W_ROOT")
+  if (is.null(r)) r <- "C:/OSGeo4W"
+  if (is.character(r) && (nchar(r) == 0)) r <- "C:/OSGeo4W"
+  # build bat file path
+  bat <- normalize_path(file.path(r, "OSGeo4W.bat"), mustWork = FALSE)
+  if (!file.exists(bat)) {
+    stop(paste("OSGeo4W not available at", dirname(bat)))
+  }
+  # find gdal_calc.py
+  p <- Sys.glob(
+    normalize_path(
+      paste0(r, "/apps/Pyth*/Scripts/gdal_calc.py"),
+      mustWork = FALSE
+    )
+  )
+  assertthat::assert_that(
+    identical(length(p), 1L),
+    msg = "could not find \"gdal_calc.py\" in OSGeo4W installation"
+  )
+  # build command
+  paste0("\"", bat, "\" \"", normalize_path(p), "\" ", x)
+}
+# nocov end
+
+python_gdal_calc <- function(x) {
+  # assert valid arguments
+  assertthat::assert_that(
+    assertthat::is.string(x),
+    assertthat::noNA(x)
+  )
+  # get Python command
+  python_cmd <- Sys.getenv("GDAL_PYTHON")
+  script_path <- Sys.getenv("GDAL_CALC")
+
+  # prepend command
+  cmd <- ""
+  if (!identical(python_cmd, "")) {
+    cmd <- paste(python_cmd, " ") # nocov
+  }
+
+  # specify script
+  if (identical(script_path, "")) {
+    script_path <- "gdal_calc.py" # nocov
+  }
+
+  # return command
+  paste0(cmd, script_path, " ", x)
 }