+++ tests/testthat/test_calc_spp_frc_data.R
@@ -80,7 +80,7 @@ test_that("different engines produce same result", {
   # skip if needed
   skip_on_cran()
   skip_if_not_installed("gdalUtilities")
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(
@@ -158,6 +158,7 @@ test_that("different engines produce same result", {
 test_that("example data", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_not_installed("vcr")
   skip_if_cached_data_not_available()
   skip_if_zenodo_data_not_available(latest_lumb_cgls_version)