+++ tests/testthat/test_create_spp_aoh_data.R
@@ -87,7 +87,7 @@ test_that("simulated data (GDAL engine)", {
   skip_on_cran()
   skip_if_not_installed("rgdal")
   skip_if_not_installed("gdalUtilities")
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(
@@ -176,6 +176,7 @@ test_that("simulated data (GRASS engine)", {
   skip_on_os("windows")
   skip_if_not_installed("rgdal")
   skip_if_not_installed("gdalUtilities")
+  skip_if_not_installed("rgrass7")
   skip_if_grass_not_available()
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
@@ -262,6 +263,7 @@ test_that("simulated data (GRASS engine)", {
 test_that("example data", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_not_installed("rgdal")
   skip_if_not_installed("vcr")
   skip_if_cached_data_not_available()
@@ -446,8 +448,9 @@ test_that("some species missing habitat data", {
 test_that("amphibian data (IUCN format)", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -512,8 +515,9 @@ test_that("amphibian data (IUCN format)", {
 test_that("reptile data (IUCN format)", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -578,8 +582,9 @@ test_that("reptile data (IUCN format)", {
 test_that("terrestrial mammal data (IUCN format)", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -644,8 +649,9 @@ test_that("terrestrial mammal data (IUCN format)", {
 test_that("bird data (BirdLife format)", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -711,8 +717,9 @@ test_that("bird data (BirdLife format)", {
 test_that("bird data (alternate BirdLife format)", {
   # skip if needed
   skip_on_cran()
+  skip_on_ci()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")