+++ tests/testthat/test_get_jung_plvl1_habitat_data.R
@@ -6,7 +6,6 @@ test_that("latest version (from online)", {
   skip_on_ci()
   skip_if_offline()
   skip_if_local_and_slow_internet()
-  skip_if_zenodo_api_not_available()
   # create object
   d <- new_temp_dir()
   x <- get_jung_plvl1_habitat_data(
@@ -32,7 +31,6 @@ test_that("latest version (from cache)", {
   skip_on_cran()
   skip_if_offline()
   skip_if_not_installed("rappdirs")
-  skip_if_zenodo_api_not_available()
   skip_if_cached_data_not_available()
   skip_if_zenodo_data_not_available(latest_jung_potential_version)
   # create object
@@ -58,7 +56,6 @@ test_that("specified version (from online)", {
   skip_on_ci()
   skip_if_offline()
   skip_if_local_and_slow_internet()
-  skip_if_zenodo_api_not_available()
   skip_if_zenodo_data_not_available(latest_jung_potential_version)
   # create object
   d <- new_temp_dir()