+++ NEWS.md
@@ -1,3 +1,13 @@
+# aoh 0.0.1.3
+
+- Fix issues with using GDAL engine for processing data on Windows.
+  Specifically, the package will now correctly work with the GDAL Python
+  bindings that are installed as part of the OSGeo4W software.
+- Fix issues where functions for downloading datasets would fail to
+  load data from cache, and needlessly re-download data.
+- Fix issue where attempting to download data or query data version
+  numbers would throw an error due to errors with the Zenodo API.
+
 # aoh 0.0.1.2
 
 - Suppress meaningless warnings triggered for GRASS on macOS.